extrac the dev pieces & consolidate stream base
